Core migration: Calculations.UnitConverter
