+++ source/Hanssens.Net.Tests/Hanssens.Net.Tests.csproj
@@ -53,7 +53,6 @@
     <Compile Include="StringExtensionTests\LongestWordTests.cs" />
     <Compile Include="LoggingTests\BaseLoggerTests.cs" />
     <Compile Include="StringExtensionTests\ToSeoFriendlyStringTests.cs" />
-    <Compile Include="CalculationsTests\UnitConverterTests.cs" />
     <Compile Include="CryptographyTests\PasswordHashTests.cs" />
     <Compile Include="ReflectionTests\ExtractPropertyValueTests.cs" />
     <Compile Include="Entities\Customer.cs" />
